+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Header V2
 
 - Adds a third radio button, separating search options between the Circulating Catalog, Research Catalog, and the NYPL.org website.
+- Temporarily removes the "log in" state, the feature to read auth cookies, and the patron's name when logged in, and the log out link.
 
 ### Adds